add function to check input parameters

+++ push-swap/turk_sort/libft/libft.h
@@ -6,17 +6,17 @@
 /*   By: ytoshihi <ytoshihi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/05 19:29:47 by ytoshihi          #+#    #+#             */
-/*   Updated: 2024/03/29 15:42:56 by ytoshihi         ###   ########.fr       */
+/*   Updated: 2024/04/02 15:43:32 by ytoshihi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_H
 # define LIBFT_H
 
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <unistd.h>
+# include <stdio.h>
+# include <stdlib.h>
+# include <string.h>
+# include <unistd.h>
 
 typedef struct s_list
 {
@@ -44,7 +44,7 @@ int			ft_strncmp(const char *s1, const char *s2, size_t n);
 void		*ft_memchr(const void *s, int c, size_t n);
 int			ft_memcmp(const void *s1, const void *s2, size_t n);
 char		*ft_strnstr(const char *haystack, const char *needle, size_t len);
-int			ft_atoi(const char *str);
+long			ft_atoi(const char *str);
 char		*ft_strdup(const char *s1);
 void		*ft_calloc(size_t num, size_t size);
 char		*ft_substr(char const *s, unsigned int start, size_t len);
